use one vector operator

+++ src/GibbsSampler.h
@@ -23,11 +23,11 @@ public:
     ColMatrix mAMeanMatrix, mAStdMatrix;
     unsigned mStatUpdates;
 
-    double mMaxGibbsMassA;
-    double mMaxGibbsMassP;
+    float mMaxGibbsMassA;
+    float mMaxGibbsMassP;
 
-    double mAnnealingTemp;
-    double mChi2;
+    float mAnnealingTemp;
+    float mChi2;
 
     bool mSingleCellRNASeq;
 
@@ -40,30 +40,30 @@ public:
     bool exchange(AtomicSupport &domain, AtomicProposal &proposal);
 
     bool evaluateChange(AtomicSupport &domain, const AtomicProposal &proposal,
-        double threshold, bool accept=false);
+        float threshold, bool accept=false);
 
-    double computeDeltaLL(const MatrixChange &change);
+    float computeDeltaLL(const MatrixChange &change);
 
-    double getGibbsMass(const MatrixChange &change);
+    float getGibbsMass(const MatrixChange &change);
 
-    void updateAPMatrix_A(unsigned row, unsigned col, double delta);
-    void updateAPMatrix_P(unsigned row, unsigned col, double delta);
+    void updateAPMatrix_A(unsigned row, unsigned col, float delta);
+    void updateAPMatrix_P(unsigned row, unsigned col, float delta);
     void updateAPMatrix(const MatrixChange &change);
 
     bool canUseGibbs(const MatrixChange &ch);
-    void setChi2(double chi2);
+    void setChi2(float chi2);
 
 public:
 
     GibbsSampler(Rcpp::NumericMatrix D, Rcpp::NumericMatrix S, unsigned nFactor,
-        double alphaA, double alphaP, double maxGibbsMassA, double maxGibbsMassP,
+        float alphaA, float alphaP, float maxGibbsMassA, float maxGibbsMassP,
         bool singleCellRNASeq, Rcpp::NumericMatrix fixedPat, char whichMat);
 
     void update(char matrixLabel);
 
     uint64_t totalNumAtoms(char matrixLabel) const;
-    void setAnnealingTemp(double temp);
-    double chi2() const;
+    void setAnnealingTemp(float temp);
+    float chi2() const;
 
     Rcpp::NumericMatrix AMeanRMatrix() const;
     Rcpp::NumericMatrix AStdRMatrix() const;
